- Made mapbox data stored globally to enable data to be reused rather than refetched each time map is loaded
