+++ src/app/app.component.html
@@ -68,5 +68,5 @@ http://docs.telerik.com/devtools/nativescript-ui/Controls/NativeScript/SideDrawe
             </StackLayout>
         </ScrollView>
     </GridLayout>
-    <page-router-outlet tkMainContent class="page page-content"></page-router-outlet>
+    <page-router-outlet tkMainContent class="page page-content" (activate)="onRouterOutletActivate($event)"></page-router-outlet>
 </RadSideDrawer>
